Validating compound values for games
If the user has provided a username assume that all of the associated
fields need to be provided as well.

This requires that each game grouping be wrapped in a <fieldset>.

This is not wired into .validate() for reporting the messages so the
console should be used when testing.

With the sample values populated for WoW and a user supplied value for
LoL (with no region) you should see the following:

  game exists:true
  index.js:30 checking compound values for ta_WOW_id
  index.js:39 compound values provided: true
  index.js:30 checking compound values for ta_LOL_id
  index.js:39 compound values provided: false
  index.js:50 provided games valid: false